SAT/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
passlib[argon2]
python-jose
